requirements.txt
django
djangorestframework
coreapi
Markdown
Pygments
django-filter
django-guardian
django-sslserver
xmltodict
psycopg2-binary